+++ NEWS.md
@@ -2,6 +2,8 @@
 * Added new exported functions `ReadExposure`, `WriteExposure`, `SortExposure`, 
 `PlotExposure`, `PlotExposureToPdf` and necessary tests and test data.
 
+* Added new internal function `NumFromId` and removed `ICAMSxtra` from the dependency packages.
+
 # 2.1.2
 * Enabled argument `target.sig.index` to also take the signature id as input in
 function `SignaturePresenceTest1`. Changed argument `m.opts` to use return value